+++ bindings/vue/src/components/index.js
@@ -1,6 +1,7 @@
 import VOnsNavigator from './VOnsNavigator.vue';
 import VOnsBackButton from './VOnsBackButton.vue';
 import VOnsTabbar from './VOnsTabbar.vue';
+import VOnsRange from './VOnsRange.vue';
 import VOnsSwitch from './VOnsSwitch.vue';
 import VOnsPullHook from './VOnsPullHook.vue';
 import VOnsSplitterSide from './VOnsSplitterSide.vue';
@@ -9,6 +10,7 @@ export {
   VOnsNavigator,
   VOnsBackButton,
   VOnsTabbar,
+  VOnsRange,
   VOnsSwitch,
   VOnsPullHook,
   VOnsSplitterSide