+++ bindings/vue/CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
 
 ### Features
  * Make `v-ons-switch` component compatible with Vue.js 2.x.
+ * Add `v-ons-range` component.
 
 <!--
 ### Bug Fixes